Add one configuration for developers to easily ignore index
This option is for OAP developers to easily analyze the OAP
performance w/ and w/o index even if the index is created already.

Usage:
sqlContext.conf.setConfString(SQLConf.OAP_USE_INDEX.key, value)

The value is true by default. This means that the index will be
used by default. Otherwise, the index won't be used.