Directly traversing row IDs in fiber caches
This patch provides the lightweight functionality to directly
traverse the row IDs in fiber caches which are compatible with
the roaring bitmap format.
It doens't need to create roaring bitmap object on heap.

Meanwhile this patch provides two test cases to test the functionality
of traversing the row IDs in single fiber cache or multi fiber caches
for two typical equal query cases with bitmap index.
One is single equal value case. The other is multi equal value case.

1. select * from t where a = apple
2. select * from t where a in {"apple", "banana", ...}

Signed-off-by: Lei Li <lei.l.li@intel.com>