Add four orc record readers
The four orc readers are below:

OrcColumnarBatchReader.java and extended
IndexedOrcColumnarBatchReader.java

OrcMapreduceRecordReader.java and extended
IndexedOrcMapreduceRecordReader.java

OrcColumnarBatchReader is a copy of OrcColumnarBatchReader
from Spark2.3 with minor changes to be able to have a child class
IndexedOrcColumnarBatchReader.

OrcMapreduceRecordReader is a copy of OrcMapreduceRecordReader
from orc upstream with minor changes to be able to have a child class
IndexedOrcMapreduceRecordReader.

The above four readers are specifically used by oap to enable orc file
format reader to use oap index.